Add files via upload
Iris is a genus of around 300 species of flowering plants. It takes the name from the Greek goddess of rainbow, Iris.

The Iris flower data set is also know as the Fisher's Iris data set. It is a multivariate data set introduced by Ronald Fisher in his paper, The use of multiple measurements in taxonomic problems as an example of linear discriminant analysis.